Move JoinForm to separate file
